V1.2
Added more location text.

+++ Shared/Utilities/MediaType.swift
@@ -21,7 +21,7 @@ enum MediaType: String, CaseIterable, Identifiable {
             return NSLocalizedString("People", comment: "")
         }
     }
-    var watchlistInt: Int {
+    var watchlistInt: Int64 {
         switch self {
         case .movie: return 0
         case .tvShow: return 1
